added teli-op from the example for Macanium Wheels
code look mostly fine, I think this is whare Justin got his code for last year
seems robost and well done. Exspet better commented and in java. I will keep
it in java for know at least untill I have the op-mode ready for a PR and with
the DC motor or servor for the shooter working with encoders.