base for running with and endoder
this uses a while loop that dose not have exit cases and this code needs to be refactored so it can chage the state
of the motor. But is works for a motor. this is acurate with in 150 ticks both ways at full power. this is from the
kick back and should not be much of an issue when we are finding the RPM of the motor.